requirements.txt
PyNaCl>=1.0.1
discord.py>=1.3.3
geocoder>=1.38.1
giphypop>=0.3
google-api-python-client>=1.9.3
google-auth-httplib2>=0.0.3
google-auth-oauthlib>=0.4.1
python-dateutil>=2.8.0
python-forecastio>=1.3.5
pyyaml>=5.3.1
